# backend/api/swagger.yaml
definitions:
...
  api.GetSet.getSetResponse:
    properties:
      set:
        $ref: '#/definitions/model.CardSet'
    type: object
  api.GetSetCards.getSetCardsResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/model.Card'
        type: array
      set:
        $ref: '#/definitions/model.CardSet'
    type: object
  api.GetSets.getSetsResponse:
    properties:
      sets:
        items:
          $ref: '#/definitions/model.CardSet'
        type: array
    type: object
...
  model.Card:
    properties:
      atk:
        type: integer
      attribute:
        type: string
      card_images:
        items:
          $ref: '#/definitions/model.CardImage'
        type: array
      card_sets:
        items:
          $ref: '#/definitions/model.CardSet'
        type: array
      def:
        type: integer
      desc:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      race:
        type: string
      sets:
        type: string
      type:
        type: string
    type: object
  model.CardImage:
    properties:
      id:
        type: integer
      image_url:
        type: string
      image_url_small:
        type: string
    type: object
  model.CardSet:
    properties:
      set_code:
        type: string
      set_name:
        type: string
      set_rarity:
        type: string
      set_rarity_code:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /sets:
    get:
      description: Retrieve all sets.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetSets.getSetsResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retrieve all sets.
      tags:
      - Yu-Gi-Oh
  /sets/{code}:
    get:
      description: Retrieve a specific set.
      parameters:
      - description: The identifier of the set.
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetSet.getSetResponse'
        "404":
          description: Set does not exist.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retrieve a specific set.
      tags:
      - Yu-Gi-Oh
  /sets/{code}/cards:
    get:
      description: Retrieve all cards of a specific set.
      parameters:
      - description: The identifier of the set.
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetSetCards.getSetCardsResponse'
        "404":
          description: Set does not exist or set contains no cards.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retrieve all cards of a specific set.
      tags:
      - Yu-Gi-Oh
...
swagger: "2.0"